Correção de palavra errada

+++ ucb.aplicacao.model/Tarefa.java
@@ -28,7 +28,7 @@ public class Tarefa {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
